sched/{core,cpufreq_schedutil}: add capacity clamping for RT/DL tasks
Currently schedutil enforce a maximum OPP when RT tasks are RUNNABLE.
Such a mandatory policy can be made more tunable from userspace thus
allowing for example to define a reasonable max frequency which is
required for the execution of a specific RT workload. This will
contribute to make the RT class more friendly for power/energy sensible
applications.

This patch extends the usage of capacity_{min,max} to the RT/DL classes.
Whenever a task in these classes in RUNNABLE, the capacity required is
defined by the constraints of the control group the task belong to.

Change-Id: I6c992ae06e59842cfee90915a20c8cce32440d4b
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org