cpufreq: schedutil: relax sg_cpus's flags on IDLE enter
Currently, sg_cpu's flags are set to the value defined by the last call of
the cpufreq_update_util()/cpufreq_update_this_cpu() and the rather simple
policy used by the RT/DL classes enforces the SCHED_CPUFREQ_{RT/DL} flags
at each call.

When multiple CPUs shares the same frequency domain it happens that a
CPU which executed an RT task, right before entering IDLE, has one of the
SCHED_CPUFREQ_RT_DL flags set, permanently, until it exit IDLE.

Thus, in sugov_next_freq_shared(), where utilisation and flags are
aggregated across all the CPUs of a frequency domain, it turns out that all
the CPUs of that domain will always run at the maximum OPP until another
event happens in the idle CPU to eventually clear the SCHED_CPUFREQ_{RT/DL}
flag.

Such a behaviour can harm the energy efficiency of systems when RT workloads
are not so frequent and other CPUs in the same frequency domain are running
small utilisation workloads, which is a quite common scenario in mobile
embedded systems.

This patch proposes a solution which is aligned with the current principle
to update the flags each time a scheduling event happens. The scheduling
of the idle_task on a CPU is considered one of such meaningful events.
That's why when the idle_task is selected for execution we poke the
schedutil policy to reset the flags for that CPU.

Moreover, no frequency transitions are activated at that point, which is
fair in case the RT workload should come back in the future, but it allows
other CPUs in the same frequency domain to scale down the frequency in
case that should be possible.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>