[NOT PATCH] Cover letter, to be removed before posting
The current version of schedutil exposes some issues related to the management
of update flags used by systems with frequency domains spawning multiple CPUs.

Each time a CPU utilisation update is issued by a scheduler a set of flags
are configured to define (mainly) which class is asking for a utilisation
update. These flags are than used by the frequency selection policy to
identify the OPP to activate.

In the current implementation, CPU flags are overridden each time the
scheduler calls schedutil for an update. Such a behaviour produces issues
in these scenarios, where we assume CPU1 and CPU2 share the same frequency
domain:
a) a RT task which executed on CPU1 can keep the domain at an high frequency
   for a long period of time, even if there are anymore RT workload in that
   domain
b) a FAIR task co-scheduled in the same CPU of a RT task can override the
   flags configured by the RT task

These misbehaves have been verified using a set of simple rt-app based
synthetic workloads, running on a ARM's Juno board, and results are
available in this Notebook [1].

This series proposes a set of fixes for the aforementioned issues as well
as a small improvement to speedup the selection of the maximum frequency
when RT tasks enter a CPU.

[1] https://gist.github.com/d6a21b459a18091b2b058668a550010d

Change-Id: I148780ba074d3cc8d327a31ae26e9b6fcab071f9