[NOT PATCH] Cover letter, to be removed before posting
To: linux-kernel@vger.kernel.org,
    linux-pm@vger.kernel.org

Cc: Ingo Molnar <mingo@redhat.com>,
    Peter Zijlstra <peterz@infradead.org>,
    Rafael J. Wysocki <rafael.j.wysocki@intel.com>,
    Viresh Kumar <viresh.kumar@linaro.org>,
    Vincent Guittot <vincent.guittot@linaro.org>,
    John Stultz <john.stultz@linaro.org>,
    Juri Lelli <juri.lelli@arm.com>,
    Todd Kjos <tkjos@android.com>,
    Tim Murray <timmurray@google.com>,
    Andres Oportus <andresoportus@google.com>,
    Joel Fernandes <joelaf@google.com>,
    Morten Rasmussen <morten.rasmussen@arm.com>,
    Dietmar Eggemann <dietmar.eggemann@arm.com>

The current version of schedutil has some issues related to the management
of update flags used by systems with frequency domains spawning multiple CPUs.

Each time a CPU utilisation update is issued by the scheduler a set of flags
are configured to define (mainly) which class is asking for a utilisation
update. These flags are then used by the frequency selection policy to
identify the OPP to choose.

In the current implementation, CPU flags are overridden each time the
scheduler calls schedutil for an update. Such a behaviour produces issues
in these scenarios, where we assume CPU1 and CPU2 share the same frequency
domain:
a) a RT task which executed on CPU1 can keep the domain at an high frequency
   for a long period of time, even if there are no longer RT tasks on
   CPUs in that domain
b) a FAIR task co-scheduled in the same CPU of a RT task can override the
   flags configured by the RT task and potentially this can cause an
   unwanted frequency drop

These misbehaviours have been verified using a set of simple rt-app based
synthetic workloads, running on a ARM's Juno board, and results are
available in this Notebook [1].

This series proposes a set of fixes for the aforementioned issues as well
as a small improvement to speedup the selection of the maximum frequency
when RT tasks enter a CPU.

[1] https://gist.github.com/d6a21b459a18091b2b058668a550010d

Change-Id: I148780ba074d3cc8d327a31ae26e9b6fcab071f9