cpufreq: schedutil: ignore the sugov kthread for frequencies selections
In system where multiple CPUs shares the same frequency domain a small
workload on a CPU can be forced to run almost always at the (not required)
highest frequency just because the RT sugov's kthread is pinned to run on
a different CPU of the same frequency domain.

Since the sugov kthread is a special RT task, which goal is just that to
activate a frequency transition, it does not make sense to have it biasing the
schedutil's frequency selection.

This patch exploits the information related to the current task to silently
ignore cpufreq_update_this_cpu() calls, coming from the RT scheduler, while
the sugov kthread is running.

Change-Id: I2b8d2c6bb28ab4df5f7165df93e9a48635a6014f
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Viresh Kumar <viresh.kumar@linaro.org>
Cc: Steve Muckle <smuckle@linaro.org>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org