cpufreq: schedutil: ensure max frequency while running RT/DL tasks
The rather simple policy in use for RT/DL tasks should set the maximum
frequency when a task in these classes calls for a cpufreq_update_this_cpu().
However, the current implementation allows to lower the frequency while a
RT/DL task is running, just because of a FAIR task waking up and being
enqueued.

This issue is due to the fact that the sg_cpu's flags are overwritten
at each call of sugov_update_*, thus the wakeup of a FAIR task resets the
flags and can trigger a frequency update which affect the currently running
RT/DL task.

This can be fixed, in multi-CPUs frequency domains, by adding (instead of
overwriting) the new flags before triggering a frequency update.
This grants to stay at least at the frequency requested by the RT/DL
class, which is the maximum one for the time being, but can also be lower
when for example DL will be extended to provide a precise bandwidth
requirement.

For single CPU frequency domains, by simply forcing the selection
of the maximum frequency in each update which happens while a RT/DL task
is running on that CPU.

Change-Id: Ibb8c8ee64a2f55a7a9cd1a3d78155bbd7560b205
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Viresh Kumar <viresh.kumar@linaro.org>
Cc: Steve Muckle <smuckle@linaro.org>
Cc: linux-kernel@vger.kernel.org
Cc: linux-pm@vger.kernel.org