[NOT PATCH] Cover letter, to be removed before posting
To: linux-kernel@vger.kernel.org,
    linux-pm@vger.kernel.org

Cc: Ingo Molnar <mingo@redhat.com>,
    Peter Zijlstra <peterz@infradead.org>,
    Tejun Heo <tj@kernel.org>,
    Rafael J. Wysocki <rafael.j.wysocki@intel.com>,
    Paul Turner <pjt@google.com>,
    Vincent Guittot <vincent.guittot@linaro.org>,
    John Stultz <john.stultz@linaro.org>,
    Morten Rasmussen <morten.rasmussen@arm.com>,
    Dietmar Eggemann <dietmar.eggemann@arm.com>,
    Juri Lelli <juri.lelli@arm.com>,
    Tim Murray <timmurray@google.com>,
    Todd Kjos <tkjos@android.com>,
    Andres Oportus <andresoportus@google.com>,
    Joel Fernandes <joelaf@google.com>

Subject: [RFC v4 0/6] Add utilization clamping support to the CPU controller

Was:
- RFCv3: Add capacity capping support to the CPU controller
- RFCv2: SchedTune: central, scheduler-driven, power-perfomance control

This is a respin of the series implementing a support for per-task boosting
and capping of CPU frequencies. This new version addresses most of the
comments collected since last posting on LKML [1] and at the OSPM Summit [2].

Hereafter a short description of the main changes.

.:: Concept: "capacity clamping" replaced by "utilization clamping"

The previous implementation was expressed in terms of "capacity
clamping", which generated some confusion mainly due to the fact that in
mainline the capacity is currently defined as a "constant property" of a
CPU. Here is a email [3] which resumes the confusion generated in by the
previous proposal.

As PeterZ pointed out, the goal of this proposal is to "affect" the
util_avg metric, i.e. the CPU utilization, and the way it's used for
example by schedutil. Thus, it actually makes a lot more sense to talk
about "utilization clamping".

In this new proposal the CPU controller exposes a couple of new attributes
to define what is the minimum and maximum utilization which should be
considered for tasks a that group. These utilization clamp values
can be used when a CPU frequency has to be selected to either "boost" or
"cap" the selected frequency.

.:: Implementation: rb-trees replaced by reference counting

The previous implementation used a couple of rb-trees to aggregate the
different clamp values of tasks co-scheduler on the same CPU.  Although
being a simple solution, from the coding standpoint, PeterZ pointed out
that it was definitively adding not negligible overheads in the fast path
(i.e. tasks enqueue/dequeue) especially on highly loaded systems.

This new implementation has been based on a mach more lightweight mechanism
using reference counting. The new solution requires just the increment
(decrement) of an integer counter each time a task is enqueue (dequeue).
The more expensive operation is now a sequential scan of a small array of
integers, which is however defined to easily fit into a single cache line.

.:: Other comments: use-cases description

People had concerns about use-cases, this [4] is a posting where I've resumed
the main use cases we are targeting with this proposal. Further discussion
went on at OSPM outside of the official tracks and I've got the feeling
that people (at least Peter and Raphael) seems to recognize the utility to
have a support to either boost or clamp the CPU frequencies based on
currently active tasks.

Main use cases was:

- boosting: better service and interactive response for small tasks which
  are affecting the user experience. Consider for example the case of a
  small control thread for an external accelerator (e.g. GPU, DSP, other
  devices).  In this case the scheduler does not have a complete view of
  what are the task bandwidth requirements and if it's small it will keep
  scheduling it at a lower frequency thus affecting the overall time
  required to complete its activation.

- clamping: increase energy efficiency for background tasks not directly
  affecting the user experience. Since running at a lower frequency is in
  general more energy efficient, when the completion time is not a main
  goal then clamping the maximum frequency to use for certain (maybe big)
  tasks can have positive effects both on power dissipation and energy
  consumption.
  Moreover, this last support allows also to make RT tasks more energy
  friendly on mobile systems, whenever running them at the maximum
  frequency is not strictly required.

.:: Other comments: usage of CGroups as a main interface

The current implementation is based on CGroups but does not rely on that
API. We do not propose a different main interface just because, so fare,
all the use-cases we have on hand can take advantage from a CGroups API.

In case there should be the need for a different API, the current
implementation can be easily extended to hock its internals such a
different API. However, it's not worth adding such an API until there is
not a real demand.

.:: Patches organization

The first three patches of this series introduces util_{min,max} tracking
in the core scheduler, as an extension of the CPU controller.
The fourth patch is dedicated to the synchronization among the cgroup
interface (slow-path) and the scheduler (fast-path).
The last two patches integrate the utilization clamping support with
schedutil for FAIR tasks and RT/DL tasks too.

Cheers Patrick

.:: References
[1] https://lkml.org/lkml/2017/2/28/355
[2] slides: http://retis.sssup.it/ospm-summit/Downloads/deadline_reclaiming.pdf
    video:  http://youtu.be/6MC1jbYbQTo
[3] https://lkml.org/lkml/2017/4/11/670
[4] https://lkml.org/lkml/2017/3/20/688